v0.68 smoothing some rough edges.  last sexification step remains to be done

+++ Source/Game/Arrow.h
@@ -72,7 +72,7 @@ public:
   long time;
   long yPos;
   float beatLength;
-  long length;
+  int length;
   int type; // for drawing based on note value
   Rating rating;
   FreezeRating freezeRating;
